+++ packages/react-billing/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.6.5](https://github.com/ttoss/ttoss/compare/@ttoss/react-billing@0.6.4...@ttoss/react-billing@0.6.5) (2026-06-12)
+
+**Note:** Version bump only for package @ttoss/react-billing
+
 ## [0.6.4](https://github.com/ttoss/ttoss/compare/@ttoss/react-billing@0.6.3...@ttoss/react-billing@0.6.4) (2026-06-10)
 
 **Note:** Version bump only for package @ttoss/react-billing